+++ src/lib/badge-tooltip/index.ts
@@ -1,0 +1,3 @@
+export * from './badge-tooltip.types';
+export * from './badge-tooltip.token';
+export * from './badge-tooltip.provider';